feat(instancing, gpu_driven): far-only mesh LODs, no GPU readback stalls; GPU-driven material work
- Instancer: a mesh LOD without surfaces hides the block instead of swapping the MultiMesh mesh, and set_mesh
  is skipped when unchanged (set_mesh after the buffer upload reads the buffer back from the GPU, ~5 ms).
  New blocks get the first non-empty mesh LOD before their buffer upload.
- Instancer: blocks of items with distance colliders keep a CPU copy of their transforms, so colliders spawn
  without get_instance_transform readbacks (4-10 ms hitches while walking). Cleared when instances are removed.
- Instancer task output: layer ids are uint16 (hashed library ids above 255 were dropped).
- GPU-driven renderer and material, engine and LOD terrain work in progress.

+++ terrain/instancing/instancer_task_output_queue.h
@@ -11,7 +11,9 @@ namespace zylann::voxel {
 
 struct InstanceLoadingTaskOutput {
 	Vector3i render_block_position;
-	uint8_t layer_id;
+	// uint16_t like every other layer id (library ids go up to 0xffff): as uint8_t, results for ids above 255 were
+	// truncated onto the wrong or a missing layer and silently dropped
+	uint16_t layer_id;
 	// Tells which parts of the block contain edited data (non-generated).
 	// When data chunks are half the size of render chunks, this is 8 bits in XYZ order.
 	uint8_t edited_mask;
